Revert "Feature/bill summary"
