+++ src/components/Layout.tsx
@@ -2,14 +2,15 @@ import styled from '@emotion/styled'
 import React, { ReactNode } from 'react'
 
 const Container = styled.div`
+  background-color: #e8e8e8;
+  height: 100vh;
   padding: 3rem 0;
-  background-color: LightGray;
 `
 
 const Content = styled.div`
   max-width: 768px;
   margin: 0 auto;
-  background-color: White;
+  background-color: white;
   padding: 1.5rem;
 `
