+++ src/constants.ts
@@ -3,5 +3,5 @@ export const BILLING_TABLE_COLUMNS = [
   '가맹점명',
   '이용금액',
   '이용구분',
-  '출금처',
+  '매입상태',
 ] as const
