+++ src/pages/Billing.tsx
@@ -1,34 +1,69 @@
-import React, { useCallback, useState } from 'react'
+import styled from '@emotion/styled'
+import React, { useRef, useState } from 'react'
+import { Button } from 'semantic-ui-react'
+import XLSX from 'xlsx'
 
 import BillingTable from '@/components/BillingTable'
-import BillSummary from '@/components/BillSummary'
 import Layout from '@/components/Layout'
 import { Bill } from '@/types'
+import { readFileAsArrayBuffer } from '@/utils/fileUtil'
+
+import { convertKBankBills, KBankBill, sheetToJson } from '../utils/xlsxUtil'
+
+const ButtonContainer = styled.div`
+  display: flex;
+  justify-content: flex-end;
+`
 
 function Billing() {
+  const inputRef = useRef<HTMLInputElement>(null)
   const [bills, setBills] = useState<Bill[]>([])
-  const [selectedBills, setSelectedBills] = useState<Bill[]>([])
 
-  const handleLoadBill = useCallback((bills: Bill[]) => {
-    setBills(bills)
-  }, [])
+  const handleLoadBillClick = () => {
+    if (!inputRef.current) {
+      return
+    }
 
-  const handleAddBill = (bill: Bill) => {
-    setBills([...bills, bill])
+    inputRef.current.click()
   }
 
-  const handleSelect = useCallback((selectedBills: Bill[]) => {
-    setSelectedBills(selectedBills)
-  }, [])
+  const handleBillFileChange = async (
+    e: React.ChangeEvent<HTMLInputElement>,
+  ) => {
+    const files = e.target.files
+    if (!files) {
+      return
+    }
+
+    try {
+      const arrayBuffer = await readFileAsArrayBuffer(files[0])
+
+      const { SheetNames, Sheets } = XLSX.read(arrayBuffer, { type: 'array' })
+      const workSheet = Sheets[SheetNames[0]]
+
+      const kBankBills = sheetToJson(workSheet) as KBankBill[]
+      const bills = convertKBankBills(kBankBills)
+
+      setBills(bills)
+    } catch {}
+  }
 
   return (
     <Layout>
-      <BillSummary
-        onLoadBill={handleLoadBill}
-        selectedBills={selectedBills}
-        onAddBill={handleAddBill}
-      />
-      <BillingTable bills={bills} onSelect={handleSelect} />
+      <ButtonContainer>
+        <Button primary size="large" onClick={handleLoadBillClick}>
+          이용 내역 불러오기
+        </Button>
+        <input
+          ref={inputRef}
+          type="file"
+          accept=".xls, .xlsx"
+          hidden
+          onChange={handleBillFileChange}
+        />
+      </ButtonContainer>
+
+      <BillingTable bills={bills} />
     </Layout>
   )
 }
